fix: move animation list
