+++ src/components/Transition/index.ts
@@ -1,3 +1,5 @@
 import Transition from './Transition'
 export default Transition
 export * from './Transition'
+
+export * from './animations'
