+++ src/components/index.ts
@@ -106,6 +106,9 @@ export * from './Tooltip'
 export { default as Callout } from './Callout'
 export * from './Callout'
 
+export { default as Transition } from './Transition'
+export * from './Transition'
+
 export * from './Filters'
 export * from './Tabs'
 export * from './FormControl'
