+++ src/lib/index.ts
@@ -4,4 +4,3 @@ export * from './primitives'
 export * from './shape'
 export * from './spacing'
 export * from './font'
-export * from './animations'
